AI/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.0
sqlalchemy[asyncio]==2.0.31
alembic==1.13.2
asyncpg==0.29.0
psycopg2-binary==2.9.9
redis==5.0.4
minio==7.2.3
qdrant-client==1.9.1
llama-index==0.9.48
llama-index-llms-gemini
llama-index-llms-ollama
llama-index-embeddings-ollama
langgraph==0.2.28
python-multipart==0.0.9
apscheduler==3.10.4
pydantic-settings==2.5.2
requests==2.32.3
python-dotenv==1.0.1
pytest==8.3.2
httpx==0.27.0
prometheus-fastapi-instrumentator==7.0.0
aiosqlite==0.20.0
google-generativeai==0.8.3
Pillow==10.4.0
PyPDF2==3.0.1
pdf2image==1.17.0
opencv-python==4.10.0.84
